agent5-reserve-calculations/requirements.txt
# Web framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# PostgreSQL
sqlalchemy==2.0.27
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# HTTP / async
httpx==0.27.0

# Data processing
numpy==1.26.4

# Testing
pytest==8.0.1
pytest-asyncio==0.23.5
pytest-httpx==0.30.0

# Date utilities
python-dateutil==2.8.2

# Utilities
python-dotenv==1.0.1
structlog==24.1.0
